requirements.txt
fpdf==1.7.2
pillow==11.3.0
